==> 02_tootle_visualiseeri_andmed.Rmd <==
---
title: "Töötle ja visualiseeri Riigikogu liikmete aktiivsuse andmed"
---

```{r}
library(tidyverse)
library(janitor)
library(lubridate)
library(runner)
library(tictoc)
library(ggrepel)
library(hrbrthemes)

artiklid_raw <- read_rds("data/artiklid_koik.rds")

sonavotud_raw <- read_rds("data/sonavotud.rds")

aruparimised_raw <- read_rds("data/aruparimised.rds")

kirjalikud_kusimused_raw <- read_rds("data/kirjalikud_kusimused.rds")

detailinfo_lingid_raw <- read_rds("data/detailinfo_lingid.rds")

fraktsioonid_raw <- read_rds("data/fraktsioonid.rds")
```

## Tootle andmed

Töötle andmed nii, et iga XIII Riigikogu liikme kohta on 1 rida.

Riigikogu liikmete nimed
```{r}
nimi <- detailinfo_lingid_raw %>% 
  distinct(riigikogu_liige) %>% 
  pull(riigikogu_liige)
```


Töötle sõnavõttude andmeid
```{r}
sonavotud <- sonavotud_raw %>% 
  mutate(kp = as.Date(value, "%d.%m.%Y")) %>% 
  filter(kp > as.Date("2015-04-01")) %>% 
  group_by(riigikogu_liige) %>% 
  summarise(sonavotte = sum(sonavotte)) %>% 
  arrange(desc(sonavotte))
```

Funktsioon, mis tuvastab iga Riigikogu liikme kohta, mitu arupärimist ta on esitanud.
```{r}
tuvasta_aruparimiste_arv <- function(x){

  aruparimised_raw %>% 
    filter(str_detect(esitaja, x)) %>% 
    summarise(aruparimisi = n_distinct(nr)) %>% 
    mutate(riigikogu_liige = x)
  
}

tuvasta_aruparimiste_arv_possibly <- possibly(tuvasta_aruparimiste_arv, NULL)
```

Tuvasta iga Riigikogu liikme kohta tema arupärimiste arv.
```{r}
aruparimised <- map_df(nimi, tuvasta_aruparimiste_arv_possibly)
```

Kirjalike küsimuste arv kõigi Riigikogu liikmete kohta
```{r}
kirjalikud_kusimused <- kirjalikud_kusimused_raw %>%
  filter(esitaja %in% nimi) %>% 
  group_by(esitaja) %>% 
  summarise(kirjalikke_kusimusi = n_distinct(nr)) %>% 
  arrange(desc(kirjalikke_kusimusi)) %>% 
  rename(riigikogu_liige = esitaja)

kirjalikud_kusimused
```

Artiklite arv kõigi Riigikogu liikmete kohta
```{r}
artiklid <- artiklid_raw %>% 
  group_by(riigikogu_liige) %>% 
  summarise(artikleid = n()) %>% 
  ungroup()
```


Pane ühte tabelisse kokku artiklite, sõnavõttude jne arv iga Riigikogu liikme kohta
```{r}
riigikogu_liikmete_aktiivsus <- detailinfo_lingid_raw %>% 
  distinct(riigikogu_liige) %>% 
  left_join(sonavotud) %>% 
  left_join(aruparimised) %>% 
  left_join(kirjalikud_kusimused) %>% 
  left_join(artiklid) %>% 
  ungroup() %>% 
  mutate_at(c("sonavotte", "aruparimisi", "kirjalikke_kusimusi", "artikleid"),
             function(x){coalesce(x, 0L)}) %>% 
  mutate(suuliselt = sonavotte,
         kirjalikult = aruparimisi + kirjalikke_kusimusi + artikleid) %>% 
  arrange(desc(suuliselt))

riigikogu_liikmete_aktiivsus
```

Salvesta koondandmed
```{r}
riigikogu_liikmete_aktiivsus %>% 
  write_rds("data/riigikogu_liikmete_aktiivsus.rds")
```



Visualiseeri aktiivsus
```{r fig.height=10, fig.width=10}
mediaan_aktiivsus <- riigikogu_liikmete_aktiivsus %>% 
  ungroup() %>% 
  summarise(suuliselt = median(suuliselt),
            kirjalikult = median(kirjalikult))

riigikogu_liikmete_aktiivsus %>% 
  left_join(fraktsioonid_raw) %>% 
  mutate(fraktsioon = case_when(riigikogu_liige == "Krista Aru" ~ "Eesti Vabaerakonna fraktsioon",
                                is.na(fraktsioon) ~ "Eesti Reformierakonna fraktsioon",
                                TRUE ~ fraktsioon),
         fraktsioon = case_when(fraktsioon == "Eesti Keskerakonna fraktsioon" ~ "Keskerakond",
                                fraktsioon == "Eesti Konservatiivse Rahvaerakonna fraktsioon" ~ "EKRE",
                                fraktsioon == "Eesti Reformierakonna fraktsioon" ~ "Reformierakond",
                                fraktsioon == "Eesti Vabaerakonna fraktsioon" ~ "Vabaerakond",
                                fraktsioon == "Fraktsiooni mittekuuluvad saadikud" ~ "Fraktsioonitud",
                                fraktsioon == "Isamaa fraktsioon" ~ "Isamaa",
                                fraktsioon == "Sotsiaaldemokraatliku Erakonna fraktsioon" ~ "Sotsid")) %>% 
  filter(!riigikogu_liige %in% c("Eiki Nestor", "Enn Eesmaa", "Helir-Valdor Seeder",
                                 "Kalle Laanet")) %>% 
  mutate(nimi = riigikogu_liige) %>% 
  separate(riigikogu_liige, into = c("eesnimi", "perenimi"), sep = " ") %>% 
  mutate(nimi = ifelse(perenimi %in% c("Aru", "Helme", "Korb"), nimi, perenimi)) %>% 
  ggplot(aes(suuliselt, kirjalikult, colour = fraktsioon)) +
  geom_point(size = 2.5) +
  geom_hline(yintercept = mediaan_aktiivsus$kirjalikult) +
  geom_vline(xintercept = mediaan_aktiivsus$suuliselt) +
  geom_text_repel(aes(label = nimi), min.segment.length = 3,
                  segment.alpha = 0.2, segment.size = 0.5,
                  show.legend = FALSE, size = 3.5) +
  scale_colour_manual(values = c("Reformierakond" = "#fec44f",
                                 "Vabaerakond" = "#295AA6",
                                 "Isamaa" = "#00AEEF",
                                 "Keskerakond" = "#007F5C",
                                 "EKRE" = "#993404",
                                 "Sotsid" = "#E10600",
                                 "Fraktsioonitud" = "#737373")) +
  annotate("text", x = mediaan_aktiivsus$suuliselt - 25, y = 550, 
           label = str_c("sõnavõttude mediaan ", mediaan_aktiivsus$suuliselt), 
           size = 4, angle = 90) +
  annotate("text", x = 1200, y = mediaan_aktiivsus$kirjalikult + 15, 
           label = str_c("kirjalik mediaan ", mediaan_aktiivsus$kirjalikult), 
           size = 4) +
  theme_ipsum_rc() +
  theme(legend.title = element_blank(),
        legend.position = "top",
        legend.text = element_text(size = 12), 
        legend.key.size = unit(1.5, "lines")) +
  guides(colour = guide_legend(override.aes = list(size = 4), nrow = 1)) +
  labs(title = "XIII Riigikogu liikmete aktiivsus",
       subtitle = "Kui palju on Riigikogu liikmed esitanud kirjalikke küsimusi ja arupärimisi \nning kirjutanud artikleid vs kui palju on esinetud sõnavõttudega?",
       y = "arupärimised + kirjalikud küsimused + artiklid",
       x = "sõnavõtud")

```


Aktiivsus fraktsioonide kaupa
```{r fig.height=10, fig.width=8}
riigikogu_liikmete_aktiivsus %>% 
  left_join(fraktsioonid_raw) %>% 
  mutate(fraktsioon = case_when(riigikogu_liige == "Krista Aru" ~ "Eesti Vabaerakonna fraktsioon",
                                is.na(fraktsioon) ~ "Eesti Reformierakonna fraktsioon",
                                TRUE ~ fraktsioon),
         fraktsioon = case_when(fraktsioon == "Eesti Keskerakonna fraktsioon" ~ "Keskerakond",
                                fraktsioon == "Eesti Konservatiivse Rahvaerakonna fraktsioon" ~ "EKRE",
                                fraktsioon == "Eesti Reformierakonna fraktsioon" ~ "Reformierakond",
                                fraktsioon == "Eesti Vabaerakonna fraktsioon" ~ "Vabaerakond",
                                fraktsioon == "Fraktsiooni mittekuuluvad saadikud" ~ "Fraktsioonitud",
                                fraktsioon == "Isamaa fraktsioon" ~ "Isamaa",
                                fraktsioon == "Sotsiaaldemokraatliku Erakonna fraktsioon" ~ "Sotsid")) %>% 
  filter(!riigikogu_liige %in% c("Eiki Nestor", "Enn Eesmaa", "Helir-Valdor Seeder",
                                 "Kalle Laanet")) %>% 
  mutate(nimi = riigikogu_liige) %>% 
  separate(riigikogu_liige, into = c("eesnimi", "perenimi"), sep = " ") %>% 
  mutate(nimi = ifelse(perenimi %in% c("Aru", "Helme", "Korb"), nimi, perenimi)) %>% 
  ggplot(aes(suuliselt, kirjalikult, colour = fraktsioon)) +
  geom_point(size = 2.5) +
  geom_hline(yintercept = mediaan_aktiivsus$kirjalikult) +
  geom_vline(xintercept = mediaan_aktiivsus$suuliselt) +
  scale_colour_manual(values = c("Reformierakond" = "#fec44f",
                                 "Vabaerakond" = "#295AA6",
                                 "Isamaa" = "#00AEEF",
                                 "Keskerakond" = "#007F5C",
                                 "EKRE" = "#993404",
                                 "Sotsid" = "#E10600",
                                 "Fraktsioonitud" = "#737373")) +
  facet_wrap(~ fraktsioon, ncol = 2) +
  theme_ipsum_rc() +
  theme(legend.title = element_blank(),
        legend.position = "top",
        legend.text = element_text(size = 12), 
        legend.key.size = unit(1.5, "lines")) +
  guides(colour = "none") +
  labs(title = "XIII Riigikogu liikmete aktiivsus fraktsioonide kaupa",
       y = "arupärimised + kirjalikud küsimused + artiklid",
       x = "sõnavõtud")
```
